Test for Station related functions while further help need.
Station.getStationPrice() is directly output to the stdout.
Although I have tried the ways to redirect stdout to string, and it does work well,
However, inside the string stream it include a java running headler of "current workspace: ..." and so on.
see "https://stackoverflow.com/questions/1119385/junit-test-for-system-out-println" for more help.